+++ killtracker/utils.py
@@ -357,6 +357,21 @@ def create_bs_button_html(
     )
 
 
+dimensions = [(12, "t"), (9, "b"), (6, "m"), (3, "k")]
+
+
+def humanize_value(value: float, precision: int = 2) -> str:
+    """returns given value in human readable and abbreviated form
+    e.g. 1234678 -> 1.23m
+    """
+    value = float(value)
+    for exponent, identifier in dimensions:
+        if value >= pow(10, exponent):
+            return f"{value / pow(10, exponent):,.{precision}f}{identifier}"
+
+    return f"{value:,.{precision}f}"
+
+
 class JsonDateTimeDecoder(json.JSONDecoder):
     def __init__(self, *args, **kwargs) -> None:
         json.JSONDecoder.__init__(