+++ testauth/settings.py
@@ -289,6 +289,3 @@ DEFAULT_FROM_EMAIL = ""
 #######################################
 # Add any custom settings below here. #
 #######################################
-
-# aa-eveuniverse
-EVEUNIVERSE_LOAD_STARGATES = True